# apps/web/openapi.json
{
  "components": {
    "schemas": {
...
      "ApiError": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "details": {
            "additionalProperties": true,
            "title": "Details",
            "type": "object"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": ["code", "message", "details"],
        "title": "ApiError",
        "type": "object"
      },
...
      "GitHistoryEntryModel": {
        "additionalProperties": false,
        "properties": {
          "author": {
            "title": "Author",
            "type": "string"
          },
          "commit": {
            "title": "Commit",
            "type": "string"
          },
          "date": {
            "title": "Date",
            "type": "string"
          },
          "path": {
            "title": "Path",
            "type": "string"
          },
          "subject": {
            "title": "Subject",
            "type": "string"
          }
        },
        "required": ["commit", "author", "date", "subject", "path"],
        "title": "GitHistoryEntryModel",
        "type": "object"
      },
...
      "GraphNodeModel": {
        "additionalProperties": false,
        "properties": {
          "category": {
            "enum": [
              "class",
              "property",
              "individual",
              "concept",
              "shape",
              "module",
              "literal",
              "resource"
            ],
            "title": "Category",
            "type": "string"
          },
          "compact": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Compact"
          },
          "datatype": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Datatype"
          },
          "kind": {
            "enum": ["iri", "bnode", "literal"],
            "title": "Kind",
            "type": "string"
          },
          "language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          },
          "module": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Module"
          },
          "value": {
            "title": "Value",
            "type": "string"
          }
        },
        "required": ["kind", "value", "category"],
        "title": "GraphNodeModel",
        "type": "object"
      },
...
      "NeighborhoodModel": {
        "additionalProperties": false,
        "properties": {
          "center": {
            "$ref": "#/components/schemas/GraphNodeModel"
          },
          "depth": {
            "title": "Depth",
            "type": "integer"
          },
          "edges": {
            "items": {
              "$ref": "#/components/schemas/QuadModel"
            },
            "title": "Edges",
            "type": "array"
          },
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/GraphNodeModel"
            },
            "title": "Nodes",
            "type": "array"
          },
          "truncated": {
            "title": "Truncated",
            "type": "boolean"
          }
        },
        "required": ["center", "depth", "nodes", "edges", "truncated"],
        "title": "NeighborhoodModel",
        "type": "object"
      },
...
      "QuadModel": {
        "additionalProperties": false,
        "properties": {
          "graph": {
            "$ref": "#/components/schemas/RdfValueModel"
          },
          "object": {
            "$ref": "#/components/schemas/RdfValueModel"
          },
          "predicate": {
            "$ref": "#/components/schemas/RdfValueModel"
          },
          "priority": {
            "default": 7,
            "maximum": 7.0,
            "minimum": 1.0,
            "title": "Priority",
            "type": "integer"
          },
          "relationship_kind": {
            "default": "other",
            "enum": [
              "subclass",
              "subproperty",
              "domain",
              "range",
              "import",
              "internal_object_property",
              "other"
            ],
            "title": "Relationship Kind",
            "type": "string"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "subject": {
            "$ref": "#/components/schemas/RdfValueModel"
          }
        },
        "required": ["subject", "predicate", "object", "graph"],
        "title": "QuadModel",
        "type": "object"
      },
      "RdfValueModel": {
        "additionalProperties": false,
        "properties": {
          "compact": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Compact"
          },
          "datatype": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Datatype"
          },
          "kind": {
            "enum": ["iri", "bnode", "literal"],
            "title": "Kind",
            "type": "string"
          },
          "language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          },
          "value": {
            "title": "Value",
            "type": "string"
          }
        },
        "required": ["kind", "value"],
        "title": "RdfValueModel",
        "type": "object"
      },
...
      "ResourceDescriptionModel": {
        "additionalProperties": false,
        "properties": {
          "definitions": {
            "items": {
              "$ref": "#/components/schemas/RdfValueModel"
            },
            "title": "Definitions",
            "type": "array"
          },
          "direct_modules": {
            "items": {
              "$ref": "#/components/schemas/RdfValueModel"
            },
            "title": "Direct Modules",
            "type": "array"
          },
          "domains": {
            "items": {
              "$ref": "#/components/schemas/RdfValueModel"
            },
            "title": "Domains",
            "type": "array"
          },
          "git_history": {
            "items": {
              "$ref": "#/components/schemas/GitHistoryEntryModel"
            },
            "title": "Git History",
            "type": "array"
          },
          "incoming": {
            "items": {
              "$ref": "#/components/schemas/QuadModel"
            },
            "title": "Incoming",
            "type": "array"
          },
          "labels": {
            "items": {
              "$ref": "#/components/schemas/RdfValueModel"
            },
            "title": "Labels",
            "type": "array"
          },
          "modules": {
            "items": {
              "$ref": "#/components/schemas/RdfValueModel"
            },
            "title": "Modules",
            "type": "array"
          },
          "outgoing": {
            "items": {
              "$ref": "#/components/schemas/QuadModel"
            },
            "title": "Outgoing",
            "type": "array"
          },
          "predicate_uses": {
            "items": {
              "$ref": "#/components/schemas/QuadModel"
            },
            "title": "Predicate Uses",
            "type": "array"
          },
          "provenance": {
            "items": {
              "$ref": "#/components/schemas/QuadModel"
            },
            "title": "Provenance",
            "type": "array"
          },
          "ranges": {
            "items": {
              "$ref": "#/components/schemas/RdfValueModel"
            },
            "title": "Ranges",
            "type": "array"
          },
          "resource": {
            "$ref": "#/components/schemas/RdfValueModel"
          },
          "shapes": {
            "items": {
              "$ref": "#/components/schemas/RdfValueModel"
            },
            "title": "Shapes",
            "type": "array"
          },
          "status": {
            "items": {
              "$ref": "#/components/schemas/RdfValueModel"
            },
            "title": "Status",
            "type": "array"
          },
          "subclasses": {
            "items": {
              "$ref": "#/components/schemas/RdfValueModel"
            },
            "title": "Subclasses",
            "type": "array"
          },
          "superclasses": {
            "items": {
              "$ref": "#/components/schemas/RdfValueModel"
            },
            "title": "Superclasses",
            "type": "array"
          },
          "types": {
            "items": {
              "$ref": "#/components/schemas/RdfValueModel"
            },
            "title": "Types",
            "type": "array"
          },
          "usage": {
            "$ref": "#/components/schemas/ResourceUsageModel"
          }
        },
        "required": [
          "resource",
          "types",
          "labels",
          "definitions",
          "modules",
          "direct_modules",
          "status",
          "outgoing",
          "incoming",
          "superclasses",
          "subclasses",
          "domains",
          "ranges",
          "shapes",
          "provenance",
          "predicate_uses",
          "usage",
          "git_history"
        ],
        "title": "ResourceDescriptionModel",
        "type": "object"
      },
      "ResourceUsageModel": {
        "additionalProperties": false,
        "properties": {
          "incoming_references": {
            "title": "Incoming References",
            "type": "integer"
          },
          "outgoing_statements": {
            "title": "Outgoing Statements",
            "type": "integer"
          },
          "predicate_uses": {
            "title": "Predicate Uses",
            "type": "integer"
          }
        },
        "required": [
          "incoming_references",
          "outgoing_statements",
          "predicate_uses"
        ],
        "title": "ResourceUsageModel",
        "type": "object"
      },
...
      "SearchConfirmationModel": {
        "additionalProperties": false,
        "properties": {
          "confirmed": {
            "title": "Confirmed",
            "type": "boolean"
          },
          "query": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Query",
            "type": "string"
          },
          "search_id": {
            "pattern": "^eow-search-v2:[A-Za-z0-9_-]+\\.[0-9a-f]{64}$",
            "title": "Search Id",
            "type": "string"
          }
        },
        "required": ["query", "confirmed", "search_id"],
        "title": "SearchConfirmationModel",
        "type": "object"
      },
      "SearchPageModel": {
        "additionalProperties": false,
        "properties": {
          "has_next": {
            "title": "Has Next",
            "type": "boolean"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/SearchResultModel"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "search_id": {
            "pattern": "^eow-search-v2:[A-Za-z0-9_-]+\\.[0-9a-f]{64}$",
            "title": "Search Id",
            "type": "string"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "offset",
          "limit",
          "has_next",
          "search_id"
        ],
        "title": "SearchPageModel",
        "type": "object"
      },
      "SearchResultModel": {
        "additionalProperties": false,
        "properties": {
          "compact_iri": {
            "title": "Compact Iri",
            "type": "string"
          },
          "iri": {
            "title": "Iri",
            "type": "string"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "local_name": {
            "title": "Local Name",
            "type": "string"
          },
          "matched_fields": {
            "items": {
              "type": "string"
            },
            "title": "Matched Fields",
            "type": "array"
          },
          "modules": {
            "items": {
              "type": "string"
            },
            "title": "Modules",
            "type": "array"
          },
          "score": {
            "title": "Score",
            "type": "integer"
          },
          "types": {
            "items": {
              "type": "string"
            },
            "title": "Types",
            "type": "array"
          }
        },
        "required": [
          "iri",
          "compact_iri",
          "local_name",
          "label",
          "types",
          "modules",
          "matched_fields",
          "score"
        ],
        "title": "SearchResultModel",
        "type": "object"
      },
...
      "TermWriteRequest": {
        "additionalProperties": false,
        "properties": {
          "acceptance_criterion_es": {
            "anyOf": [
              {
                "maxLength": 5000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Acceptance Criterion Es"
          },
          "alternative_labels_es": {
            "items": {
              "type": "string"
            },
            "maxItems": 100,
            "title": "Alternative Labels Es",
            "type": "array"
          },
          "author": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Author",
            "type": "string"
          },
          "definition_es": {
            "maxLength": 5000,
            "minLength": 1,
            "title": "Definition Es",
            "type": "string"
          },
          "domain": {
            "anyOf": [
              {
                "maxLength": 2048,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Domain"
          },
          "evidence": {
            "maxLength": 5000,
            "minLength": 1,
            "title": "Evidence",
            "type": "string"
          },
          "form_values": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                },
                {
                  "type": "number"
                },
                {
                  "type": "boolean"
                },
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                }
              ]
            },
            "maxProperties": 100,
            "title": "Form Values",
            "type": "object"
          },
          "iri": {
            "maxLength": 2048,
            "minLength": 1,
            "title": "Iri",
            "type": "string"
          },
          "kind": {
            "enum": [
              "ontology",
              "class",
              "concept",
              "object_property",
              "datatype_property",
              "annotation_property",
              "node_shape",
              "competency_question"
            ],
            "title": "Kind",
            "type": "string"
          },
          "module_id": {
            "pattern": "^[a-z][a-z0-9_-]{0,63}$",
            "title": "Module Id",
            "type": "string"
          },
          "preferred_label_es": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Preferred Label Es",
            "type": "string"
          },
          "question_text_es": {
            "anyOf": [
              {
                "maxLength": 5000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Question Text Es"
          },
          "range": {
            "anyOf": [
              {
                "maxLength": 2048,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Range"
          },
          "reading_direction_es": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reading Direction Es"
          },
          "search": {
            "$ref": "#/components/schemas/SearchConfirmationModel"
          },
          "status": {
            "default": "proposed",
            "enum": ["proposed", "active", "deprecated"],
            "title": "Status",
            "type": "string"
          },
          "valid_example": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Valid Example"
          }
        },
        "required": [
          "iri",
          "module_id",
          "kind",
          "preferred_label_es",
          "definition_es",
          "evidence",
          "author",
          "search"
        ],
        "title": "TermWriteRequest",
        "type": "object"
      },
...
      "WriteResultModel": {
        "additionalProperties": false,
        "properties": {
          "operation": {
            "title": "Operation",
            "type": "string"
          },
          "path": {
            "title": "Path",
            "type": "string"
          },
          "preserved_unknown_triples": {
            "title": "Preserved Unknown Triples",
            "type": "integer"
          },
          "resource": {
            "title": "Resource",
            "type": "string"
          }
        },
        "required": [
          "operation",
          "resource",
          "path",
          "preserved_unknown_triples"
        ],
        "title": "WriteResultModel",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Stable API over the canonical RDF/Git workspace, including controlled proposal authoring, semantic review and the synchronized canonical agent contract.",
    "title": "Enterprise Ontology Workbench API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/graph/neighborhood": {
      "get": {
        "operationId": "neighborhood_api_graph_neighborhood_get",
        "parameters": [
          {
            "in": "query",
            "name": "center",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "Center",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "depth",
            "required": false,
            "schema": {
              "default": 1,
              "maximum": 3,
              "minimum": 0,
              "title": "Depth",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "max_nodes",
            "required": false,
            "schema": {
              "default": 500,
              "maximum": 500,
              "minimum": 1,
              "title": "Max Nodes",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "max_edges",
            "required": false,
            "schema": {
              "default": 1500,
              "maximum": 1500,
              "minimum": 0,
              "title": "Max Edges",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "graph_iri",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Graph Iri"
            }
          },
          {
            "in": "query",
            "name": "predicate",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Predicate"
            }
          },
          {
            "in": "query",
            "name": "type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Type"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NeighborhoodModel"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured API error"
          }
        },
        "summary": "Neighborhood",
        "tags": ["resources", "graph"]
      }
    },
...
    "/api/resources": {
      "patch": {
        "operationId": "write_term_api_resources_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TermWriteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WriteResultModel"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured API error"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured API error"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured API error"
          }
        },
        "summary": "Write Term",
        "tags": ["editing"]
      },
      "post": {
        "operationId": "write_term_api_resources_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TermWriteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WriteResultModel"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured API error"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured API error"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured API error"
          }
        },
        "summary": "Write Term",
        "tags": ["editing"]
      }
    },
...
    "/api/resources/describe": {
      "get": {
        "operationId": "describe_resource_api_resources_describe_get",
        "parameters": [
          {
            "in": "query",
            "name": "iri",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "Iri",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResourceDescriptionModel"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured API error"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured API error"
          }
        },
        "summary": "Describe Resource",
        "tags": ["resources"]
      }
    },
...
    "/api/resources/search": {
      "get": {
        "operationId": "search_resources_api_resources_search_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Type"
            }
          },
          {
            "in": "query",
            "name": "module",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Module"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchPageModel"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured API error"
          }
        },
        "summary": "Search Resources",
        "tags": ["resources"]
      }
    },
...
  }
}